rename all files from camelCase to clearer pattern
