+++ services/authors/src/app.js
@@ -1,6 +1,6 @@
 import express from 'express';
 import bodyParser from 'body-parser';
-import authorRouter from './routers/authorRouter.js';
+import authorRouter from './routers/author.router.js';
 import { PREFIX } from './config.js';
 
 const app = express();
