+++ services/authors/src/controllers/author.controller.js
@@ -1,9 +1,9 @@
-import authorServices from '../services/authorServices.js';
+import authorService from '../services/author.service.js';
 
 class AuthorController {
   async getAuthors(req, res) {
     try {
-      const authors = await authorServices.getAuthors();
+      const authors = await authorService.getAuthors();
       res.status(200).json(authors);
     } catch (err) {
       res.status(404).json({ message: err.message });
@@ -13,7 +13,7 @@ class AuthorController {
   async getAuthorById(req, res) {
     try {
       const { id } = req.params;
-      const author = await authorServices.getAuthorById(id);
+      const author = await authorService.getAuthorById(id);
       res.status(200).json(author);
     } catch (err) {
       console.log(err);
@@ -24,7 +24,7 @@ class AuthorController {
   async createAuthor(req, res) {
     try {
       const authorData = req.body;
-      const createdAuthor = await authorServices.createAuthor(authorData);
+      const createdAuthor = await authorService.createAuthor(authorData);
       res.status(201).json(createdAuthor);
     } catch (err) {
       res.status(400).json({ message: err.message });
@@ -35,7 +35,7 @@ class AuthorController {
     try {
       const { id } = req.params;
       const authorData = req.body;
-      const updatedAuthor = await authorServices.updateAuthor(id, authorData);
+      const updatedAuthor = await authorService.updateAuthor(id, authorData);
       res.status(200).json(updatedAuthor);
     } catch (err) {
       res.status(404).json({ message: err.message });
@@ -45,7 +45,7 @@ class AuthorController {
   async deleteAuthor(req, res) {
     try {
       const { id } = req.params;
-      const deletedAuthor = await authorServices.deleteAuthor(id);
+      const deletedAuthor = await authorService.deleteAuthor(id);
       res.status(200).json(deletedAuthor);
     } catch (err) {
       res.status(404).json({ message: err.message });
